ML-1407 merge D365 workbasket + case summary into lcml.view.details
Fold the workbasket and case-summary assertions into the existing
lcml.view.details D365 flow instead of a separate feature, reusing the
submission Given and D365 helpers. Replace the no-op Internal (D365) View
details scenario (Application URL is absent for marine licence cases, so
it asserted nothing) with two meaningful scenarios: the case shown in the
workbasket (columns + row) and the Case summary tab details. Delete the
standalone workbasket feature/steps files.
