+++ test/features/lcml/lcml.view.details.feature
@@ -16,8 +16,20 @@ Feature: LCML: View details page shows sites and activities
     Then the View details page shows the "File upload" site location method
     And the View details page shows an uploaded site card with a name and a map
 
-  @real-defra-id @d365
-  Scenario: Internal (D365) View details shows the submitted sites and activities
+  @real-defra-id @d365 @issue=ML-1407
+  Scenario: A submitted marine licence case is shown in the D365 workbasket
     Given an organisation user has submitted a marine licence application with uploaded sites
-    When the marine licence case is opened from its D365 Application URL
-    Then the internal View details page shows the submitted sites and activities
+    When the internal user finds the submitted case in the Marine licence cases workbasket
+    Then the Marine licence cases workbasket displays the following columns
+      | Reference       |
+      | Project name    |
+      | Assigned to     |
+      | Status          |
+      | Case age (days) |
+    And the workbasket row shows the submitted case reference, project name and status "Submitted"
+
+  @real-defra-id @d365 @issue=ML-1407
+  Scenario: The D365 Case summary tab displays the submitted marine licence case details
+    Given an organisation user has submitted a marine licence application with uploaded sites
+    When the internal user opens the submitted case summary in D365
+    Then the case summary displays the marine licence case details
